state save fixes; ui fixes and logic fixes

+++ app/src/main/java/com/ainullov/kamil/transportation_problem/presentation/ui/solution/graph/GraphAdapter.kt
@@ -11,15 +11,14 @@ import de.blox.graphview.Graph
 class GraphAdapter(
     graph: Graph,
     private val graphForAdapter: Graph = graph,
-    private val onClickListener: (Int) -> Unit,
-    private val onLongClickListener: (Int) -> Unit): BaseGraphAdapter<GraphViewHolder>(graph) {
+    private val onClickListener: () -> Unit,
+    private val onLongClickListener: () -> Unit): BaseGraphAdapter<GraphViewHolder>(graph) {
     override fun onCreateViewHolder(parent: ViewGroup?, viewType: Int): GraphViewHolder {
         val view: View = LayoutInflater.from(parent!!.context).inflate(R.layout.item_node, parent, false)
         return GraphViewHolder(view)
     }
 
     override fun onBindViewHolder(viewHolder: GraphViewHolder, data: Any?, position: Int) {
-//        (viewHolder as GraphViewHolder).textView.setText(attr.data.toString())
         viewHolder.bind(graph = graphForAdapter,
             position = position,
             onClickListener = onClickListener,
